Add initial support for Settings - this allows the user to set their sort prefs
Also adds a new page in the settings application, as well as the ability
to sort by last name (in addition to first name).

Due to a known issue, the user must restart contacts (if currently running)
to switch to their new sort preferences.  This issue will be resolved in
a future patch.

Signed-off-by: Kaitlin Rupert <kaitlin.rupert@intel.com>