Re-add changes to disable save in New view (until user enters valid data)
Support for this was dropped when the FooterBar component was added.  This also
makes it so that if the user strips all data from a contact in edit view, save
is disabled.  That way a user can't enter an invalid contact from Edit or New.

Also, remove the check in createPerson() for valid data. This check is now
handled on the QML side.

Signed-off-by: Kaitlin Rupert <kaitlin.rupert@intel.com>