BMC#15831 - Make sure that we don't do a dataReset() after adding a contact
Adding a new contact results in a duplicate contact in the list view.  This
occurs due to the following flow:

1) A synchronous save is done in createPersonModel() - this immediately saves
   the contact to tracker
2) The view shifts from NewContactViewPortrait to GroupedViewPortrait, which
   in turn calls setFilter(AllFilter) to be called.
3) setFilter(AllFilter) fetches all of the contacts from tracker - this
   includes the newly added contact.  This updates the view.
4) Saving the contact to tracker emits a signal addContact, which is being
   handled asynchronously.  The addContact signal handler then adds the new
   contact to the model and the view is updated.

Unfortunately, the addContact signal is usually handled after the setFilter()
call has already occured.  So we end up with the new contact in the model twice,
but only in tracker once.

The fix here is really a hack.  Filtering should probably be done in the
proxymodel, but searchContacts() relies on updating the filter.  So
restructuring is needed to move the filtering to the proxy model.  Also, this
issue will likely disappear (or at least change somewhat) when all calls to the
QContactManager are moved to asynchronous calls.

Signed-off-by: Kaitlin Rupert <kaitlin.rupert@intel.com>