BMC#18432: Make DropDown width in ContactsSettings dependent on width of parent
Instead of setting a static size for the width of DropDown, place the DropDown
in an Item that is anchored to the right side of settingsText and to the right
of the parent.  The DropDown component will set the width of the DropDown to
the width of the parent by default.

Since the DropDown component doesn't check to see if the ActionMenu it launches
will extend beyond the size of the parent, set the maxWidth of the parent to
be half of the parent.

Signed-off-by: Kaitlin Rupert <kaitlin.rupert@intel.com>