BMC#18559: Ensure index bar characters don't display beyond the visible window
Place the ListView in an Item which is anchored to the right side of the
divIcon and to the right side of the window.  Each delegate in the ListView
can then be of width parent.width / count.  This ensures that the items fit
within the available space in the FooterBar.

Additionally, if the application is in portrait mode, reduce the number of
characters visible if the number of characters exceeds 10.  This will help
ensure that the characters are not displayed too close together.

Signed-off-by: Kaitlin Rupert <kaitlin.rupert@intel.com>