Add a second street field in address details
This also fixes a bug where the wrong values would show up in a given address
field in Edit View. For example, if your address was:

Street
City
State

In Edit View, you might see City in the state field and State in the zip field.
This is because we were stripping out the "\n" character in the AddressRole so
that the Detail View wouldn't show empty fields.  However, the "\n" is used
to denote which value belongs to which field in the Edit View.

Signed-off-by: Kaitlin Rupert <kaitlin.rupert@intel.com>