BMC#20241,BMC#20243: Fix issues with adding IM details to contact, online status
Due to changes in meego-ux-components, the DropDowns in ImEditWidget were no
longer working properly.  This patch makes changes to ensure the IM details
are correctly added to the contact.

This also fixes issues with the online status - a colored status indicator dot
should always be present on all contacts.  This fix also ensures that the
accountsModel from telepathy is available before we assign the contact status.

Signed-off-by: Kaitlin Rupert <kaitlin.rupert@intel.com>