Remove parentTitle properties from pages
This doesn't need to be passed to the pages, as we can query the window for the
page title using:

  window.pageStack.currentPage.pageTitle

Also, don't attempt to restore data until after page is done loading.  This
ensures that the page title is set correctly and that nothing else is in flight
due to the page change.

Calling restoreData() in finishPageLoad() can probably be used instead of the
Timer{} in ContactsExpandable.  Will follow up in another patch to fix this.

Signed-off-by: Kaitlin Rupert <kaitlin.rupert@intel.com>