Replace prefixSaveRestore with pageTitle
prefixSaveRestore was always getting set to an empty string.  This fix replaces
the empty string with the page title.  This ensures that values from the New
page don't clobber those from the Edit page.  There shouldn't be a need to
restore both sets of data, but it does make it easier to clean up the key/value
pairs once we're done with them (which will follow in a later patch).

Signed-off-by: Kaitlin Rupert <kaitlin.rupert@intel.com>