Set index property of Detail/Edit to source based off currentContactIndex
On a restore, the index will be set to an invalid index when the app first
loads (as no index has been selected).  As part of the restore process, the
last known value of currentContactIndex is fetched - this needs to be passed
to the page during page activation.

Also, be sure to call restoreOnce() using SaveRestoreState object.  And use
editViewPortrait id instead of newContact.

Signed-off-by: Kaitlin Rupert <kaitlin.rupert@intel.com>