BMC#20769: Use Text element's parent leftMargin in AddressEditWidget for width
The Text element itself doesn't set a leftMargin, so we need to use the
parent's here.  Otherwise, the text field will overlap the trash can after
the user has added the first address detail.

Only subtract deleted item from height if there is more than 1 item in the list:
When deleting an item detail from a contact, items are removed from the UI
unless it's the last item.  The last item is simply cleared, not removed.

Also, remove some obsolete code the throws a QML error (which prevents items
from displaying in the UI properly after a delete has occurred).

Signed-off-by: Kaitlin Rupert <kaitlin.rupert@intel.com>