Wait until page has activated until setting the index
As the page is activating, the model hasn't finished reseting.  Ideally, we need
a signal that indicates when the model is ready to read from.  Once the model is
ready, we can then restore the page.  Need to investigate whether that'll
too much lag between app start up and page restore. This is a temporary fix in
the meantime.

Signed-off-by: Kaitlin Rupert <kaitlin.rupert@intel.com>