BMC#21111: Use the width of the window in case DropDown's parent had no width
When loading an item via a Loader, there's no easy way to set the parent of the
item until after it is fully loaded. There's a window of time where the
DropDown's parent might not be set or might have an incorrect width.  When this
happens, use the width of the window.  We want the DropDown to have a max
width of roughly half the window anyway.

Signed-off-by: Kaitlin Rupert <kaitlin.rupert@intel.com>