Set background of footer to be white - prevent a black background from showing
When clicking on a button, the background behind the button might flash
through.  This looks more natural is the background is set to white to match
the rest of the app.

Also, in the Detail view, the parent.height is the same as the flickable area -
which is only as large as the content area,  If a given contact only has first/
last/company set, there will be no other details.  This will lead to a black
page.  Set the height of the rectangle to be that of the window so that the
whole page is white.

Signed-off-by: Kaitlin Rupert <kaitlin.rupert@intel.com>