[qmlcontacts] Implement PHOTO handling in vcardconverter
Previously, PHOTO handling was left to the default implementation.
In Qt5Pim, PHOTO details are not supported (as there is no
QContactThumbnail detail in Qt5Pim).

This commit ensures that any PHOTO property encountered is saved to
disk, and then an appropriate QContactAvatar detail is generated which
points to it.